Re-enable tab entry and ensure the cursor is moved the correct width
Previously the `insert-tab` binding was a no-op because DISABLE_TABS
is `false` causing insert_tab() not to be run. Remove the guard around
 insert_tab() to ensure that a tab (or spaces) are outputted and
correct the calculation of the distance to move horizontally to work in
both the tab and pseudo-tab (spaces) cases.

Tested with DISABLE_TABS set to `true` (i.e. insert spaces) and `false`.