Add a configuration parameter for trimming whitespace on save
Prior to this change trailing whitespace is always remove when a file is saved. This can be expensive for very large files and in some cases may not be wanted by the user. Add a new configuration parameter `trim_trailing_whitespace_on_save` to allow the user to change the behaviour globally.